Refactor play to lessons
As per the new mocks, the paly tab was renamed to lessons.
Hence, the entrire project was refactored, along with tests, to maintain
consistency with mocks.

Resolves #603
