+++ app/src/main/java/org/oppia/app/fragment/FragmentComponent.kt
@@ -22,7 +22,7 @@ import org.oppia.app.testing.BindableAdapterTestFragment
 import org.oppia.app.topic.TopicFragment
 import org.oppia.app.topic.conceptcard.ConceptCardFragment
 import org.oppia.app.topic.info.TopicInfoFragment
-import org.oppia.app.topic.play.TopicPlayFragment
+import org.oppia.app.topic.lessons.TopicPlayFragment
 import org.oppia.app.topic.questionplayer.QuestionPlayerFragment
 import org.oppia.app.topic.review.TopicReviewFragment
 import org.oppia.app.topic.train.TopicTrainFragment
