+++ app/src/main/java/org/oppia/app/topic/TopicFragmentPresenter.kt
@@ -34,7 +34,7 @@ class TopicFragmentPresenter @Inject constructor(
   private val tabIcons =
     intArrayOf(
       R.drawable.ic_info_icon_24dp,
-      R.drawable.ic_play_icon_24dp,
+      R.drawable.ic_lessons_icon_24dp,
       R.drawable.ic_train_icon_24dp,
       R.drawable.ic_review_icon_24dp
     )
@@ -65,11 +65,11 @@ class TopicFragmentPresenter @Inject constructor(
     viewPager.adapter = adapter
     tabLayout.setupWithViewPager(viewPager)
     tabLayout.getTabAt(0)!!.setText(fragment.getString(R.string.info)).setIcon(tabIcons[0])
-    tabLayout.getTabAt(1)!!.setText(fragment.getString(R.string.play)).setIcon(tabIcons[1])
+    tabLayout.getTabAt(1)!!.setText(fragment.getString(R.string.lessons)).setIcon(tabIcons[1])
     tabLayout.getTabAt(2)!!.setText(fragment.getString(R.string.train)).setIcon(tabIcons[2])
     tabLayout.getTabAt(3)!!.setText(fragment.getString(R.string.review)).setIcon(tabIcons[3])
     if (topicId.isNotEmpty() && storyId.isNotEmpty())
-      setCurrentTab(TopicTab.PLAY)
+      setCurrentTab(TopicTab.LESSONS)
   }
 
   private val topicLiveData: LiveData<Topic> by lazy { getTopic() }
