+++ app/src/main/java/org/oppia/app/topic/ViewPagerAdapter.kt
@@ -5,7 +5,7 @@ import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentManager
 import androidx.fragment.app.FragmentStatePagerAdapter
 import org.oppia.app.topic.info.TopicInfoFragment
-import org.oppia.app.topic.play.TopicPlayFragment
+import org.oppia.app.topic.lessons.TopicPlayFragment
 import org.oppia.app.topic.review.TopicReviewFragment
 import org.oppia.app.topic.train.TopicTrainFragment
 
@@ -26,7 +26,7 @@ class ViewPagerAdapter(
         topicInfoTab.arguments = args
         return topicInfoTab
       }
-      TopicTab.PLAY -> {
+      TopicTab.LESSONS -> {
         val topicPlayTab = TopicPlayFragment()
         if (storyId.isNotEmpty())
           args.putString(STORY_ID_ARGUMENT_KEY, storyId)
