+++ app/src/main/java/org/oppia/app/topic/lessons/ChapterSummarySelector.kt
@@ -1,8 +1,8 @@
-package org.oppia.app.topic.play
+package org.oppia.app.topic.lessons
 
 import org.oppia.app.model.ChapterSummary
 
-/** Interface to transfer the selected chapter summary to [TopicPlayFragmentPresenter]. */
+/** Interface to transfer the selected chapter summary to [TopicLessonsFragmentPresenter]. */
 interface ChapterSummarySelector {
   fun selectChapterSummary(chapterSummary: ChapterSummary)
 }
