+++ app/src/main/java/org/oppia/app/topic/lessons/ExpandedChapterListIndexListener.kt
@@ -1,4 +1,4 @@
-package org.oppia.app.topic.play
+package org.oppia.app.topic.lessons
 
 /**
  * Interface to keep track of story-index for which ChapterList is visible/expanded.
