+++ app/src/main/java/org/oppia/app/topic/lessons/StorySummarySelector.kt
@@ -1,8 +1,8 @@
-package org.oppia.app.topic.play
+package org.oppia.app.topic.lessons
 
 import org.oppia.app.model.StorySummary
 
-/** Interface to transfer the selected story summary to [TopicPlayFragmentPresenter]. */
+/** Interface to transfer the selected story summary to [TopicLessonsFragmentPresenter]. */
 interface StorySummarySelector {
   fun selectStorySummary(storySummary: StorySummary)
 }
