+++ app/src/main/java/org/oppia/app/topic/lessons/TopicLessonsItemViewModel.kt
@@ -1,6 +1,6 @@
-package org.oppia.app.topic.play
+package org.oppia.app.topic.lessons
 
 import org.oppia.app.viewmodel.ObservableViewModel
 
 /** The root [ViewModel] for all individual items that may be displayed in [TopicPlayFragment] recycler view. */
-abstract class TopicPlayItemViewModel: ObservableViewModel()
+abstract class TopicLessonsItemViewModel: ObservableViewModel()
