+++ app/src/main/java/org/oppia/app/topic/lessons/TopicLessonsTitleViewModel.kt
@@ -1,0 +1,6 @@
+package org.oppia.app.topic.lessons
+
+import androidx.lifecycle.ViewModel
+
+/** [ViewModel] for title in [TopicPlayFragment]. */
+class TopicLessonsTitleViewModel : TopicLessonsItemViewModel()
