--- app/src/main/java/org/oppia/app/topic/play/TopicPlayTitleViewModel.kt
@@ -1,6 +1,0 @@
-package org.oppia.app.topic.play
-
-import androidx.lifecycle.ViewModel
-
-/** [ViewModel] for title in [TopicPlayFragment]. */
-class TopicPlayTitleViewModel : TopicPlayItemViewModel()
