+++ app/src/main/res/layout/topic_play_story_summary.xml
@@ -16,7 +16,7 @@
 
     <variable
       name="viewModel"
-      type="org.oppia.app.topic.play.StorySummaryViewModel" />
+      type="org.oppia.app.topic.lessons.StorySummaryViewModel" />
   </data>
 
   <com.google.android.material.card.MaterialCardView
