+++ app/src/main/res/values/strings.xml
@@ -42,7 +42,7 @@
   <string name="state_learn_again_button">Learn Again</string>
   <string name="topic_download_text">(%s)</string>
   <string name="info">Info</string>
-  <string name="play">Play</string>
+  <string name="lessons">Lessons</string>
   <string name="train">Train</string>
   <string name="review">Review</string>
   <string name="topic">Topic</string>
